refactor: dir structure changes and function signature
