--- filters/sigv4/signer/doc.go
@@ -1,1 +1,0 @@
-package signer
